+++ src/privacy_unit_tracking/privacy_unit.rs
@@ -7,7 +7,7 @@ pub const PRIVACY_COLUMNS: usize = 2;
 pub const PRIVACY_UNIT: &str = "_PRIVACY_UNIT_";
 pub const PRIVACY_UNIT_DEFAULT: &str = "_PRIVACY_UNIT_DEFAULT_";
 pub const PRIVACY_UNIT_WEIGHT: &str = "_PRIVACY_UNIT_WEIGHT_";
-pub const PER_ROW_PROTECTION: &str = "_RANDOM_";
+pub const PER_ROW_PRIVACY: &str = "_RANDOM_";
 
 // A few utility objects
 
@@ -270,8 +270,8 @@ impl PrivacyUnit {
         PRIVACY_COLUMNS
     }
 
-    pub fn per_row_protection() -> &'static str {
-        PER_ROW_PROTECTION
+    pub fn per_row_privacy() -> &'static str {
+        PER_ROW_PRIVACY
     }
 
     pub fn privacy_unit() -> &'static str {
@@ -416,7 +416,7 @@ mod tests {
             ),
             ("order_table", vec![("user_id", "user_table", "id")], "name"),
             ("user_table", vec![], "name"),
-            ("product_table", vec![], PER_ROW_PROTECTION),
+            ("product_table", vec![], PER_ROW_PRIVACY),
         ]);
         println!("{}", privacy_unit);
     }